+++ kaparoo/utils/__init__.py
@@ -17,6 +17,8 @@ __all__ = (
     "Timer",
     "UnweightedReduction",
     "Var",
+    "ensure_in_range",
+    "ensure_literal",
     "factory_if_none",
     "replace_if_none",
     "unwrap_or_default",
@@ -42,6 +44,7 @@ from kaparoo.utils.aggregate import (
     UnweightedReduction,
     Var,
 )
+from kaparoo.utils.checks import ensure_in_range, ensure_literal
 from kaparoo.utils.optional import (
     factory_if_none,
     replace_if_none,